rename and organize files
