+++ src/utils/index.ts
@@ -1,5 +1,5 @@
 export * from "./exampleCode";
-export * from "./formatting";
+export * from "./formatter";
 export * from "./parser";
 export * from "./parseObj";
 export * from "./transformers";
